+++ _pages/08-permissions.md
@@ -3,8 +3,8 @@ layout: post
 title:  "Permissions and Types"
 ---
 <script type="module">
-import {LayeredFilesystem} from "{{ '/js/lfs.js' | relative_url }}"
-import {Shell} from "{{ '/js/shell.js' | relative_url }}"
+import {LayeredFilesystem} from "{{ '/js/lfs.mjs' | relative_url }}"
+import {Shell} from "{{ '/js/shell.mjs' | relative_url }}"
 
 var shell = new Shell(new LayeredFilesystem(), document.getElementById("shell"));
 shell.main("{{ site.baseurl }}");
