+++ js/coreutils/cat.js
@@ -1,4 +1,4 @@
-import {Shell} from '../shell.js'
+import {Shell} from '../shell.mjs'
 
 Shell.prototype.handle_cat = async function(command) {
     var prog_name = command.arguments.shift();
