+++ js/coreutils/exec.js
@@ -1,16 +1,16 @@
-import {Shell} from '../shell.js'
+import {Shell} from '../shell.mjs'
 
-import {CONSTANTS} from '../defs.js'
-import {bytes_to_str} from '../fs_helper.js'
+import {CONSTANTS} from '../defs.mjs'
+import {bytes_to_str} from '../fs_helper.mjs'
 
 // Import everything so programs exec'd from the filesystem have access to it by
 // default
-import * as fs_helper from '../fs_helper.js'
-import * as defs from '../defs.js'
-import * as fs from '../fs.js'
-import * as lfs from '../lfs.js'
-import * as myfs from '../myfs.js'
-import * as shell from '../shell.js'
+import * as fs_helper from '../fs_helper.mjs'
+import * as defs from '../defs.mjs'
+import * as fs from '../fs.mjs'
+import * as lfs from '../lfs.mjs'
+import * as myfs from '../myfs.mjs'
+import * as shell from '../shell.mjs'
 
 Shell.prototype.handle_exec = async function(command) {
     command.arguments.shift();
