+++ js/coreutils/link.js
@@ -1,6 +1,6 @@
-import {Shell} from '../shell.js'
+import {Shell} from '../shell.mjs'
 
-import {str_to_bytes} from '../fs_helper.js'
+import {str_to_bytes} from '../fs_helper.mjs'
 
 Shell.prototype.handle_link = async function(command) {
     async function usage() {
