+++ js/coreutils/stat.js
@@ -1,6 +1,6 @@
-import {Shell} from '../shell.js'
+import {Shell} from '../shell.mjs'
 
-import {str_to_bytes} from '../fs_helper.js'
+import {str_to_bytes} from '../fs_helper.mjs'
 
 Shell.prototype.handle_stat = async function(command) {
     if (command.arguments.length != 2) {
