+++ spec/javascripts/test/shellUISpec.js
@@ -7,15 +7,15 @@ function TestEnv(container, fs, shell) {
 describe("Test Shell UI/input features" , function() {
     async function imports() {
         await _import({
-            src: '/__src__/js/lfs.js',
+            src: '/__src__/js/lfs.mjs',
             name: 'LayeredFilesystem'});
 
         await _import({
-            src: '/__src__/js/shell.js',
+            src: '/__src__/js/shell.mjs',
             name: 'Shell'});
 
         await _import({
-            src: '/__src__/js/fs_helper.js',
+            src: '/__src__/js/fs_helper.mjs',
             name: 'bytes_to_str'});
     }
